Refactor main init into separate package; turn off debug in deployment - refactor init code for containers into separate package for readability and (eventual) unit testing - consolidate error handling to single return/exit point from main - remove debug from deployment and wait4webhook script
