+++ platform-operator/main.go
@@ -6,54 +6,31 @@ package main
 import (
 	"flag"
 	"os"
-	"sync"
-	"time"
-
-	"github.com/verrazzano/verrazzano/pkg/constants"
-	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/webhooks"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/registry"
-	vzstatus "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/status"
-	"k8s.io/client-go/dynamic"
-	"k8s.io/client-go/rest"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-	"sigs.k8s.io/controller-runtime/pkg/webhook"
 
 	oam "github.com/crossplane/oam-kubernetes-runtime/apis/core"
 	cmapiv1 "github.com/jetstack/cert-manager/pkg/apis/certmanager/v1"
 	promoperapi "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
-	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/validators"
-	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	apiextensionsv1client "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-
 	vmov1 "github.com/verrazzano/verrazzano-monitoring-operator/pkg/apis/vmcontroller/v1"
 	vzappclusters "github.com/verrazzano/verrazzano/application-operator/apis/clusters/v1alpha1"
 	vzapp "github.com/verrazzano/verrazzano/application-operator/apis/oam/v1alpha1"
-
-	"go.uber.org/zap"
-	istioclinet "istio.io/client-go/pkg/apis/networking/v1alpha3"
-	istioclisec "istio.io/client-go/pkg/apis/security/v1beta1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/kubernetes"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
-	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	kzap "sigs.k8s.io/controller-runtime/pkg/log/zap"
-
 	"github.com/verrazzano/verrazzano/pkg/helm"
 	vzlog "github.com/verrazzano/verrazzano/pkg/log"
 	clustersv1alpha1 "github.com/verrazzano/verrazzano/platform-operator/apis/clusters/v1alpha1"
 	installv1alpha1 "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	installv1beta1 "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
-	clusterscontroller "github.com/verrazzano/verrazzano/platform-operator/controllers/clusters"
-	configmapcontroller "github.com/verrazzano/verrazzano/platform-operator/controllers/configmaps"
-	secretscontroller "github.com/verrazzano/verrazzano/platform-operator/controllers/secrets"
-	vzcontroller "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano"
+	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/validators"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/registry"
 	internalconfig "github.com/verrazzano/verrazzano/platform-operator/internal/config"
-	"github.com/verrazzano/verrazzano/platform-operator/internal/k8s/certificate"
-	"github.com/verrazzano/verrazzano/platform-operator/internal/k8s/netpolicy"
-	"github.com/verrazzano/verrazzano/platform-operator/metricsexporter"
+	operatorinit "github.com/verrazzano/verrazzano/platform-operator/internal/init"
+	"go.uber.org/zap"
+	istioclinet "istio.io/client-go/pkg/apis/networking/v1alpha3"
+	istioclisec "istio.io/client-go/pkg/apis/security/v1beta1"
+	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
+	kzap "sigs.k8s.io/controller-runtime/pkg/log/zap"
 	// +kubebuilder:scaffold:imports
 )
 
@@ -110,8 +87,8 @@ func main() {
 		"Specify the root directory of Verrazzano (used for development)")
 	flag.StringVar(&bomOverride, "bom-path", "", "BOM file location")
 	flag.BoolVar(&helm.Debug, "helm-debug", helm.Debug, "Add the --debug flag to helm commands")
-	// Set to 0 to disable the health check routine.
-	flag.Int64Var(&healthCheckPeriodSeconds, "health-check-period", 60, "Health check period seconds")
+	flag.Int64Var(&config.HealthCheckPeriodSeconds, "health-check-period", config.HealthCheckPeriodSeconds,
+		"Health check period seconds; set to 0 to disable health checks")
 
 	// Add the zap logger flag set to the CLI.
 	opts := kzap.Options{}
@@ -146,268 +123,16 @@ func main() {
 
 	registry.InitRegistry()
 	//This allows separation of webhooks and operator
+	var exitErr error
 	if config.RunWebhookInit {
-		webhookInit(config, log)
+		exitErr = operatorinit.WebhookInit(config, log)
 	} else if config.WebhooksEnabled {
-		startWebhookServers(config, log)
+		exitErr = operatorinit.StartWebhookServers(config, log, scheme)
 	} else {
-		reconcilePlatformOperator(config, log)
+		exitErr = operatorinit.ReconcilePlatformOperator(config, log, scheme)
 	}
-}
-
-// webhookInit Runs the webhook init container logic
-// - create Secrets for WebhookServers to use
-func webhookInit(config internalconfig.OperatorConfig, log *zap.SugaredLogger) {
-	log.Debug("Creating certificates used by webhooks")
-
-	conf, err := ctrl.GetConfig()
-	if err != nil {
-		log.Errorf("Failed to get kubeconfig: %v", err)
-		os.Exit(1)
-	}
-
-	kubeClient, err := kubernetes.NewForConfig(conf)
-	if err != nil {
-		log.Errorf("Failed to get clientset: %v", err)
-		os.Exit(1)
-	}
-
-	err = certificate.CreateWebhookCertificates(log, kubeClient, config.CertDir)
-	if err != nil {
-		log.Errorf("Failed to create certificates used by webhooks: %v", err)
-		os.Exit(1)
-	}
-
-}
-
-func updateWebhooks(log *zap.SugaredLogger, mgr manager.Manager, certsDir string) error {
-	log.Infof("Start called for pod %s", os.Getenv("HOSTNAME"))
-	conf, err := ctrl.GetConfig()
-	if err != nil {
-		log.Errorf("Failed to get kubeconfig: %v", err)
-		os.Exit(1)
-	}
-
-	kubeClient, err := kubernetes.NewForConfig(conf)
-	if err != nil {
-		log.Errorf("Failed to get clientset: %v", err)
-		os.Exit(1)
-	}
-
-	dynamicClient, err := dynamic.NewForConfig(conf)
-	if err != nil {
-		log.Errorf("Failed to create Kubernetes dynamic client: %v", err)
-		os.Exit(1)
-	}
-
-	updateWebhookConfigurations(kubeClient, log, conf)
-
-	createOrUpdateNetworkPolicies(conf, log, kubeClient)
-
-	setupWebhooksWithManager(log, mgr, kubeClient, dynamicClient)
-
-	mgr.GetWebhookServer().CertDir = certsDir
-
-	return nil
-}
-
-// startWebhookServers uses the same image as operator, but configured only to run Webhooks for the platform.
-func startWebhookServers(config internalconfig.OperatorConfig, log *zap.SugaredLogger) {
-	log.Debug("Creating certificates used by webhooks")
-
-	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
-		Scheme:                  scheme,
-		MetricsBindAddress:      config.MetricsAddr,
-		Port:                    9443,
-		LeaderElection:          config.LeaderElectionEnabled,
-		LeaderElectionNamespace: constants.VerrazzanoInstallNamespace,
-		LeaderElectionID:        "3ec4d295.verrazzano.io",
-	})
-	if err != nil {
-		log.Errorf("Error creating controller runtime manager: %v", err)
-		os.Exit(1)
-	}
-
-	updateWebhooks(log, mgr, config.CertDir)
-
-	// +kubebuilder:scaffold:builder
-	log.Info("Starting webhook controller-runtime manager")
-	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {
-		log.Errorf("Failed starting webhook controller-runtime manager: %v", err)
-		os.Exit(1)
-	}
-
-}
-
-func setupWebhooksWithManager(log *zap.SugaredLogger, mgr manager.Manager, kubeClient *kubernetes.Clientset, dynamicClient dynamic.Interface) {
-	// Setup the validation webhook
-	log.Debug("Setting up Verrazzano webhook with manager")
-	if err := (&installv1alpha1.Verrazzano{}).SetupWebhookWithManager(mgr, log); err != nil {
-		log.Errorf("Failed to setup install.v1alpha1.Verrazzano webhook with manager: %v", err)
-		os.Exit(1)
-	}
-	if err := (&installv1beta1.Verrazzano{}).SetupWebhookWithManager(mgr, log); err != nil {
-		log.Errorf("Failed to setup install.v1beta1.Verrazzano webhook with manager: %v", err)
-		os.Exit(1)
-	}
-
-	// register MySQL backup job mutating webhook
-	mgr.GetWebhookServer().Register(
-		constants.MysqlBackupMutatingWebhookPath,
-		&webhook.Admission{
-			Handler: &webhooks.MySQLBackupJobWebhook{
-				Client:        mgr.GetClient(),
-				KubeClient:    kubeClient,
-				DynamicClient: dynamicClient,
-				Defaulters:    []webhooks.MySQLDefaulter{},
-			},
-		},
-	)
-	// register MySQL install values webhooks
-	mgr.GetWebhookServer().Register(webhooks.MysqlInstallValuesV1beta1path, &webhook.Admission{Handler: &webhooks.MysqlValuesValidatorV1beta1{}})
-	mgr.GetWebhookServer().Register(webhooks.MysqlInstallValuesV1alpha1path, &webhook.Admission{Handler: &webhooks.MysqlValuesValidatorV1alpha1{}})
-
-	// Set up the validation webhook for VMC
-	log.Debug("Setting up VerrazzanoManagedCluster webhook with manager")
-	if err := (&clustersv1alpha1.VerrazzanoManagedCluster{}).SetupWebhookWithManager(mgr); err != nil {
-		log.Errorf("Failed to setup webhook with manager: %v", err)
-		os.Exit(1)
-	}
-}
-
-func updateWebhookConfigurations(kubeClient *kubernetes.Clientset, log *zap.SugaredLogger, conf *rest.Config) {
-	log.Debug("Delete old VPO webhook configuration")
-	if err := certificate.DeleteValidatingWebhookConfiguration(kubeClient, certificate.OldOperatorName); err != nil {
-		log.Errorf("Failed to delete old webhook configuration: %v", err)
-		os.Exit(1)
-	}
-
-	log.Debug("Updating VPO webhook configuration")
-
-	if err := certificate.UpdateValidatingWebhookConfiguration(kubeClient, certificate.OperatorName); err != nil {
-		log.Errorf("Failed to update validation webhook configuration: %v", err)
-		os.Exit(1)
-	}
-
-	log.Debug("Updating conversion webhook")
-	apixClient, err := apiextensionsv1client.NewForConfig(conf)
-	if err != nil {
-		log.Errorf("Failed to get apix clientset: %v", err)
-		os.Exit(1)
-	}
-
-	if err := certificate.UpdateConversionWebhookConfiguration(apixClient, kubeClient); err != nil {
-		log.Errorf("Failed to update conversion webhook: %v", err)
-		os.Exit(1)
-	}
-
-	if err := certificate.UpdateMutatingWebhookConfiguration(kubeClient, constants.MysqlBackupMutatingWebhookName); err != nil {
-		log.Errorf("Failed to update pod mutating webhook configuration: %v", err)
-		os.Exit(1)
-	}
-
-	log.Debug("Updating MySQL install values webhook configuration")
-	if err := certificate.UpdateValidatingWebhookConfiguration(kubeClient, webhooks.MysqlInstallValuesWebhook); err != nil {
-		log.Errorf("Failed to update validation webhook configuration: %v", err)
-		os.Exit(1)
-	}
-}
-
-func createOrUpdateNetworkPolicies(conf *rest.Config, log *zap.SugaredLogger, kubeClient *kubernetes.Clientset) {
-	c, err := client.New(conf, client.Options{})
-	if err != nil {
-		log.Errorf("Failed to get controller-runtime client: %v", err)
-		os.Exit(1)
-	}
-
-	log.Debug("Creating or updating network policies")
-	var errors []error
-	_, errors = netpolicy.CreateOrUpdateNetworkPolicies(kubeClient, c)
-	if len(errors) < 0 {
-		log.Errorf("Failed to create or update network policies: %v", err)
-		os.Exit(1)
-	}
-	if err != nil {
-		log.Errorf("Failed to create a controller-runtime manager: %v", err)
-		os.Exit(1)
-	}
-}
-
-// reconcilePlatformOperator runs the Verrazzano Platform Operator without running webhooks servers
-func reconcilePlatformOperator(config internalconfig.OperatorConfig, log *zap.SugaredLogger) {
-	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
-		Scheme:             scheme,
-		MetricsBindAddress: config.MetricsAddr,
-		Port:               8080,
-		LeaderElection:     config.LeaderElectionEnabled,
-		LeaderElectionID:   "3ec4d290.verrazzano.io",
-	})
-	if err != nil {
-		log.Errorf("Failed to create a controller-runtime manager: %v", err)
-		os.Exit(1)
-	}
-
-	metricsexporter.StartMetricsServer(log)
-
-	// Set up the reconciler
-	statusUpdater := vzstatus.NewStatusUpdater(mgr.GetClient())
-	healthCheck := vzstatus.NewHealthChecker(statusUpdater, mgr.GetClient(), time.Duration(healthCheckPeriodSeconds)*time.Second)
-	reconciler := vzcontroller.Reconciler{
-		Client:            mgr.GetClient(),
-		Scheme:            mgr.GetScheme(),
-		DryRun:            config.DryRun,
-		WatchedComponents: map[string]bool{},
-		WatchMutex:        &sync.RWMutex{},
-		StatusUpdater:     statusUpdater,
-	}
-	if err = reconciler.SetupWithManager(mgr); err != nil {
-		log.Error(err, "Failed to setup controller", vzlog.FieldController, "Verrazzano")
-		os.Exit(1)
-	}
-	if healthCheckPeriodSeconds > 0 {
-		healthCheck.Start()
-	}
-
-	// Set up the reconciler for VerrazzanoManagedCluster objects
-	if err = (&clusterscontroller.VerrazzanoManagedClusterReconciler{
-		Client: mgr.GetClient(),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr); err != nil {
-		log.Error(err, "Failed to setup controller", vzlog.FieldController, "VerrazzanoManagedCluster")
-		os.Exit(1)
-	}
-
-	// Start the goroutine to sync Rancher clusters and VerrazzanoManagedCluster objects
-	rancherClusterSyncer := &clusterscontroller.RancherClusterSyncer{
-		Client: mgr.GetClient(),
-		Log:    log,
-	}
-	go rancherClusterSyncer.StartSyncing()
-
-	// Setup secrets reconciler
-	if err = (&secretscontroller.VerrazzanoSecretsReconciler{
-		Client:        mgr.GetClient(),
-		Scheme:        mgr.GetScheme(),
-		StatusUpdater: statusUpdater,
-	}).SetupWithManager(mgr); err != nil {
-		log.Error(err, "Failed to setup controller", vzlog.FieldController, "VerrazzanoSecrets")
-		os.Exit(1)
-	}
-
-	// Setup configMaps reconciler
-	if err = (&configmapcontroller.VerrazzanoConfigMapsReconciler{
-		Client:        mgr.GetClient(),
-		Scheme:        mgr.GetScheme(),
-		StatusUpdater: statusUpdater,
-	}).SetupWithManager(mgr); err != nil {
-		log.Error(err, "Failed to setup controller", vzlog.FieldController, "VerrazzanoConfigMaps")
-		os.Exit(1)
-	}
-
-	// +kubebuilder:scaffold:builder
-	log.Info("Starting controller-runtime manager")
-	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {
-		log.Errorf("Failed starting controller-runtime manager: %v", err)
+	if exitErr != nil {
+		log.Errorf("Error occurred during execution: %v", exitErr)
 		os.Exit(1)
 	}
 }
