+++ platform-operator/scripts/wait4webhook.sh
@@ -2,7 +2,6 @@
 # Copyright (c) 2020, 2022, Oracle and/or its affiliates.
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 #While loop for verrazzano-platform-operator to wait for webhooks to be started before starting up
-set -x
 while [[ "$(curl --insecure -s -o /dev/null -w '%{http_code}' https://verrazzano-platform-operator-webhook:443/validate-install-verrazzano-io-v1alpha1-verrazzano -H 'Content-Type: application/json')" != "200" ]]; do sleep 5; done
 while [[ "$(curl --insecure -s -o /dev/null -w '%{http_code}' https://verrazzano-platform-operator-webhook:443/validate-install-verrazzano-io-v1beta1-verrazzano -H 'Content-Type: application/json')" != "200" ]]; do sleep 5; done
 while [[ "$(curl --insecure -s -o /dev/null -w '%{http_code}' https://verrazzano-platform-operator-webhook:443/validate-clusters-verrazzano-io-v1alpha1-verrazzanomanagedcluster -H 'Content-Type: application/json')" != "200" ]]; do sleep 5; done
